requirements.txt
Django==3.1.3
python-dotenv==0.15.0
djangorestframework==3.12.2
